Develop

+++ src/games/games.service.ts
@@ -17,14 +17,26 @@ export class GamesService {
   }
 
   async findAll(): Promise<Game[]> {
-    return await this.gamesRepository.find({ relations: { author: true } });
+    return await this.gamesRepository.find({
+      relations: {
+        categories: true,
+        author: true,
+        base: { categories: true, author: true },
+        variations: { categories: true, author: true },
+      }
+    });
   }
 
   async findOne(id: string): Promise<Game> {
     try {
       return await this.gamesRepository.findOneOrFail({
         where: { id: id },
-        relations: { author: true }
+        relations: {
+          categories: true,
+          author: true,
+          base: { categories: true, author: true },
+          variations: { categories: true, author: true },
+        }
       });
     } catch {
       throw new NotFoundException();
@@ -35,32 +47,30 @@ export class GamesService {
     if (updateGameDto.title && await this.gamesRepository.existsBy({ title: updateGameDto.title }))
       throw new ConflictException('Title already assigned');
 
-    try {
-      const game: Game = await this.gamesRepository.findOneOrFail({
-        where: { id: id },
-        relations: { author: true }
-      });
+    const game: Game | null = await this.gamesRepository.findOne({
+      where: { id: id },
+      relations: {
+        categories: true,
+        author: true,
+        base: { categories: true, author: true }
+      }
+    });
 
-      if (userId !== game.author.id) throw new UnauthorizedException();
+    if (!game) throw new NotFoundException();
+    if (userId !== game.author.id) throw new UnauthorizedException();
 
-      return await this.gamesRepository.save({ ...game, ...updateGameDto });
-    } catch {
-      throw new NotFoundException();
-    }
+    return await this.gamesRepository.save({ ...game, ...updateGameDto });
   }
 
   async remove(userId: string, id: string): Promise<Game> {
-    try {
-      const game: Game = await this.gamesRepository.findOneOrFail({
-        where: { id: id },
-        relations: { author: true }
-      });
+    const game: Game | null = await this.gamesRepository.findOne({
+      where: { id: id },
+      relations: { author: true, variations: { author: true } }
+    });
 
-      if (userId !== game.author.id) throw new UnauthorizedException();
+    if (!game) throw new NotFoundException();
+    if (userId !== game.author.id) throw new UnauthorizedException();
 
-      return await this.gamesRepository.remove(game);
-    } catch {
-      throw new NotFoundException();
-    }
+    return await this.gamesRepository.remove(game);
   }
 }
